Delay before first attempt to connect to rpc server
The first attempt usually comes before the server is ready. This moves
the delay from after a failed attempt to connect to befare each attempt.
This change was made because without it the benchmark runner would
almost always print an warning indicating a failure to connect the first
time it tries to connect.